// src/payload-types.ts
export interface Media {
  id: number;
  alt: string;
  caption?: string | null;
  updatedAt: string;
  createdAt: string;
  url?: string | null;
  thumbnailURL?: string | null;
  filename?: string | null;
  mimeType?: string | null;
  filesize?: number | null;
  width?: number | null;
  height?: number | null;
  focalX?: number | null;
  focalY?: number | null;
}
/**
 * This interface was referenced by `Config`'s JSON-Schema
 * via the `definition` "hero-slides".
 */
export interface HeroSlide {
  id: number;
  eyebrow?: string | null;
  heading: string;
  description: string;
  image: number | Media;
  ctaLabel?: string | null;
  ctaLink?: string | null;
  order: number;
  active: boolean;
  updatedAt: string;
  createdAt: string;
}